prac 7 - 1.1

+++ Prac 7/models.py
@@ -5,7 +5,7 @@ from sqlalchemy.exc import IntegrityError
 class Product(db.Model):
     id = db.Column(db.Integer, primary_key = True)
     name = db.Column(db.String(50), unique = True, nullable = False)
-    price = db.Column(db.Integer, nullable = False)
+    price = db.Column(db.Float(2), nullable = False)
 
 db.create_all()
